refactor: update import paths to use new directory structure
