+++ playground/src/features/enforcement/enforcement.js
@@ -1,5 +1,5 @@
 import React, {useEffect, useState} from 'react'
-import YamlEditor from "@pkg/editor/yaml";
+import YamlEditor from "@lib/editor/yaml";
 import "allotment/dist/style.css";
 import {useShapeStore} from "@state/shape";
 import yaml, {dump} from 'js-yaml';
