+++ playground/src/lib/editor/yaml/schema.json
